feat(server): 修改上传逻辑

+++ server/src/main/java/com/pighand/aio/service/common/impl/AssetsServiceImpl.java
@@ -2,15 +2,26 @@ package com.pighand.aio.service.common.impl;
 
 import com.mybatisflex.core.query.QueryWrapper;
 import com.mybatisflex.core.update.UpdateChain;
+import com.pighand.aio.common.sdk.tencentCloud.TencentCloudSDK;
+import com.pighand.aio.domain.base.ApplicationPlatformKeyDomain;
 import com.pighand.aio.domain.common.AssetsDomain;
 import com.pighand.aio.mapper.common.AssetsMapper;
+import com.pighand.aio.service.base.ApplicationPlatformKeyService;
 import com.pighand.aio.service.common.AssetsService;
+import com.pighand.aio.service.common.UploadService;
 import com.pighand.aio.vo.common.AssetsVO;
 import com.pighand.framework.spring.base.BaseServiceImpl;
 import com.pighand.framework.spring.page.PageOrList;
 import com.pighand.framework.spring.util.VerifyUtils;
+import com.qcloud.cos.model.DeleteObjectsRequest;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.pighand.aio.domain.base.table.ApplicationTableDef.APPLICATION;
+import static com.pighand.aio.domain.base.table.UserExtensionTableDef.USER_EXTENSION;
 import static com.pighand.aio.domain.common.table.AssetsTableDef.ASSETS;
 
 /**
@@ -20,17 +31,40 @@ import static com.pighand.aio.domain.common.table.AssetsTableDef.ASSETS;
  * @createDate 2024-12-31 19:04:50
  */
 @Service
+@RequiredArgsConstructor
 public class AssetsServiceImpl extends BaseServiceImpl<AssetsMapper, AssetsDomain> implements AssetsService {
 
+    private final UploadService uploadService;
+    private final ApplicationPlatformKeyService projectPlatformKeyService;
+    private final TencentCloudSDK tencentCloudSDK;
+
     /**
-     * 创建
+     * 创建。支持同时上传多个文件
      *
      * @param comAssetsVO
      * @return
      */
     @Override
     public AssetsVO create(AssetsVO comAssetsVO) {
-        super.mapper.insert(comAssetsVO);
+        List<String> urls = comAssetsVO.getUrls();
+        if (urls.size() == 1) {
+            String url = urls.get(0);
+            uploadService.updateFileOfficial(url);
+
+            comAssetsVO.setUrl(url);
+            super.mapper.insert(comAssetsVO);
+        } else {
+            List<AssetsDomain> assets = new ArrayList<>(urls.size());
+            for (String url : urls) {
+                uploadService.updateFileOfficial(url);
+
+                AssetsDomain assetsDomain = new AssetsDomain();
+                assetsDomain.setTag(comAssetsVO.getTag());
+                assetsDomain.setUrl(url);
+                assets.add(assetsDomain);
+            }
+            super.mapper.insertBatch(assets);
+        }
 
         return comAssetsVO;
     }
@@ -43,7 +77,7 @@ public class AssetsServiceImpl extends BaseServiceImpl<AssetsMapper, AssetsDomai
      */
     @Override
     public AssetsDomain find(Long id) {
-        return super.mapper.find(id, null);
+        return super.mapper.find(id);
     }
 
     /**
@@ -54,8 +88,9 @@ public class AssetsServiceImpl extends BaseServiceImpl<AssetsMapper, AssetsDomai
      */
     @Override
     public PageOrList<AssetsVO> query(AssetsVO comAssetsVO) {
+        comAssetsVO.setJoinTables(APPLICATION.getName(), USER_EXTENSION.getName());
 
-        QueryWrapper queryWrapper = QueryWrapper.create()
+        QueryWrapper queryWrapper = QueryWrapper.create().select(ASSETS.DEFAULT_COLUMNS)
 
             // like
             .and(ASSETS.TAG.like(comAssetsVO.getTag())).and(ASSETS.URL.like(comAssetsVO.getUrl()));
@@ -72,7 +107,7 @@ public class AssetsServiceImpl extends BaseServiceImpl<AssetsMapper, AssetsDomai
     public void update(AssetsVO comAssetsVO) {
         UpdateChain updateChain = this.updateChain().where(ASSETS.ID.eq(comAssetsVO.getId()));
 
-        updateChain.set(ASSETS.ID, comAssetsVO.getId(), VerifyUtils::isNotEmpty);
+        updateChain.set(ASSETS.TAG, comAssetsVO.getTag(), VerifyUtils::isNotEmpty);
 
         updateChain.update();
     }
@@ -84,6 +119,53 @@ public class AssetsServiceImpl extends BaseServiceImpl<AssetsMapper, AssetsDomai
      */
     @Override
     public void delete(Long id) {
+        AssetsDomain assets = super.queryChain().select(ASSETS.URL).where(ASSETS.ID.eq(id)).one();
+
+        if (assets == null) {
+            return;
+        }
+
+        // 删除COS文件
+        ApplicationPlatformKeyDomain applicationPlatformKey = projectPlatformKeyService.uploadKey();
+        String key = uploadService.fileKeyByUrl(assets.getUrl());
+
+        tencentCloudSDK.cosClient(applicationPlatformKey.getAppid(), applicationPlatformKey.getSecret(),
+            applicationPlatformKey.getRegion()).deleteObject(applicationPlatformKey.getBucket(), key);
+
         super.mapper.deleteById(id);
+    }
+
+    /**
+     * 批量删除
+     *
+     * @param ids
+     */
+    @Override
+    public void batchDelete(List<Long> ids) {
+        List<AssetsDomain> assets = super.queryChain().select(ASSETS.ID, ASSETS.URL).where(ASSETS.ID.in(ids)).list();
+
+        List<DeleteObjectsRequest.KeyVersion> deleteKeys = new ArrayList<>(assets.size());
+        List<Long> deleteIds = new ArrayList<>(assets.size());
+
+        // 删除COS文件
+        for (AssetsDomain assetsDomain : assets) {
+            String key = uploadService.fileKeyByUrl(assetsDomain.getUrl());
+            DeleteObjectsRequest.KeyVersion keyVersion = new DeleteObjectsRequest.KeyVersion(key);
+            deleteKeys.add(keyVersion);
+
+            deleteIds.add(assetsDomain.getId());
+        }
+
+        ApplicationPlatformKeyDomain applicationPlatformKey = projectPlatformKeyService.uploadKey();
+
+        DeleteObjectsRequest deleteObjectsRequest = new DeleteObjectsRequest(applicationPlatformKey.getBucket());
+        deleteObjectsRequest.setKeys(deleteKeys);
+        deleteObjectsRequest.setKeys(deleteKeys);
+
+        tencentCloudSDK.cosClient(applicationPlatformKey.getAppid(), applicationPlatformKey.getSecret(),
+            applicationPlatformKey.getRegion()).deleteObjects(deleteObjectsRequest);
+
+        // 删除数据库
+        super.mapper.deleteBatchByIds(deleteIds);
     }
 }
